refactor(cloud): extract Plugin logic into abstract BasePlugin in cloud-ops
